+++ NEWS.md
@@ -16,7 +16,7 @@ The `tibble()` and `as_tibble()` functions, and the low-level `new_tibble()` con
 
 - `as_tibble()` always checks that all columns are 1D or 2D vectors and not of type `POSIXlt`, even with `validate = FALSE` (which is now deprecated).
 
-- Calling `as_tibble()` on a vector now returns a one-row tibble, for consistency with `as_tibble.list()`.  Use `enframe(name = NULL)` for converting a vector to a one-column tibble.
+- Calling `as_tibble()` on a vector now warns once per session.  Use `enframe(name = NULL)` for converting a vector to a one-column tibble.
 
 - `data_frame()` and `frame_data()` are soft-deprecated, please use `tibble()` or `tribble()` (#111).